C問題
浮動小数点誤差に注意

+++ abc/169/c.py
@@ -1,6 +1,7 @@
 import math
+from decimal import Decimal
 
 A, B = input().split()
 a = int(A)
-b = float(B)
+b = Decimal(B)
 print(math.floor(a * b))
